% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{children_to_df}
\alias{children_to_df}
\title{Funcion que transforma un documento XML con una estructura padre-hijos a un tibble}
\usage{
children_to_df(xml_doc)
}
\arguments{
\item{xml_doc}{}
}
\value{
Un objeto de clase "tibble"
}
\description{
Parsea una relación de "un solo piso", de un solo padre a múltiples hijos
}
