% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diputados.R
\name{dipuActivosToDF}
\alias{dipuActivosToDF}
\title{Parsea documento XML con datos de Diputados activos a tibble}
\usage{
dipuActivosToDF(dipu_activos)
}
\arguments{
\item{dipu_activos}{documento-XML con datos de Diputados activos}
}
\description{
Parsea documento XML con datos de Diputados activos a tibble
}
