% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diputados.R
\name{dipuDatosToDF}
\alias{dipuDatosToDF}
\title{Parsea documento XML con datos de un Diputado a tibble}
\usage{
dipuDatosToDF(datos_legislador)
}
\arguments{
\item{datos_legislador}{documento-XML con datos de Diputados activos}
}
\description{
Parsea documento XML con datos de un Diputado a tibble
}
