% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diputados.R
\name{dipuHistoricoToDF}
\alias{dipuHistoricoToDF}
\title{Parsea documento XML con datos con listado historico de Diputados a tibble}
\usage{
dipuHistoricoToDF(dipu_historico)
}
\arguments{
\item{dipu_historico}{documento-XML con datos de Diputados activos}
}
\description{
Parsea documento XML con datos con listado historico de Diputados a tibble
}
