% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expedientes.R
\name{expCabezaToDF}
\alias{expCabezaToDF}
\title{Función que parsea el expediente cabeza de un expediente de XML a tibble}
\usage{
expCabezaToDF(ExpCabeza)
}
\arguments{
\item{ExpCabeza}{expediente cabeza de un expediente en formato XML}
}
\description{
Función que parsea el expediente cabeza de un expediente de XML a tibble
}
