% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expedientes.R
\name{expedienteToDF}
\alias{expedienteToDF}
\title{Función que parsea un expediente}
\usage{
expedienteToDF(Expediente = expediente)
}
\arguments{
\item{Expediente}{expediente, Ver getExpediente}
}
\description{
Función que parsea un expediente
}
