% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fix_timestamp}
\alias{fix_timestamp}
\title{Parsea las columnas de fecha-hora a POSIXct}
\usage{
fix_timestamp(ts)
}
\arguments{
\item{character}{}
}
\value{
Vector de clase POSIXct
}
\description{
Esta función elimina los puntos en "a.m." y "p.m." que impiden interpretar el vector como POSIXct y lo transforma a esa clase
}
