% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diputados.R
\name{getDiputadoDatos}
\alias{getDiputadoDatos}
\title{Trae los datos personales y partidarios por diputado}
\usage{
getDiputadoDatos(id_legislador)
}
\arguments{
\item{id_legislador}{integer (Requerido) Identificador del legislador}
}
\value{
Objeto de clase "tibble"
}
\description{
Trae los datos personales y partidarios por diputado
}
\examples{
prueba_dipuDatos <- getDiputadoDatos(30650)

}
\author{
Martin Olmos, \email{molmos@itba.edu.ar}
}
