% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diputados.R
\name{getDiputadosActivos}
\alias{getDiputadosActivos}
\title{Trae los Diputados con mandato actual}
\usage{
getDiputadosActivos(id_bloque = "")
}
\arguments{
\item{id_bloque}{integer (opcional) ID de un bloque parlamentario. En caso de estar vacio trae todos los bloques.}
}
\value{
Objeto de clase "tibble"
}
\description{
Trae los Diputados con mandato actual
}
\examples{
prueba_dipuActivos <- getDiputadosActivos(82)

}
\author{
Martin Olmos, \email{molmos@itba.edu.ar}
}
