% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diputados.R
\name{getDiputadosHistorico}
\alias{getDiputadosHistorico}
\title{Trae el listado historico de legisladores y autores (Pueden no ser legisladores)}
\usage{
getDiputadosHistorico()
}
\value{
Objeto de clase "tibble"
}
\description{
Trae el listado historico de legisladores y autores (Pueden no ser legisladores)
}
\examples{
prueba_dipuHistorico <- getDiputadosHistorico(30650)

}
\author{
Martin Olmos, \email{molmos@itba.edu.ar}
}
