% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expedientes.R
\name{getExpediente}
\alias{getExpediente}
\title{Función que hace un requerimiento de expedientes}
\usage{
getExpediente(IdProyectoTipo = "", IdAutoresInternos = "",
  IdUbicacion = "", IdEstado = "", Sumario = "",
  SumarioExacto = "0", FechaDesde = "", FechaHasta = "",
  AnioParlamentario = "", Limite = "")
}
\arguments{
\item{IdProyectoTipo}{integer, ID del tipo de proyecto. Ver getTiposProyecto}

\item{IdAutoresInternos}{integer, ID de el/la autor/a # debería haber un lugar donde buscar el id por nombre}

\item{IdUbicacion}{integer, ID de la ubicación}

\item{IdEstado}{integer, ID del Estado}

\item{Sumario}{character, Texto a buscar en el sumario}

\item{SumarioExacto}{integer, Macheo exacto: 0 para Falso y 1 para Verdadero}

\item{FechaDesde}{character, Acotar búsqueda desde fecha. Formato dd/mm/aaaa}

\item{FechaHasta}{character, Acotar búqueda hasta fecha. Formato dd/mm/aaaa}

\item{AnioParlamentario}{integer, Acotar la búsqueda a un año. Formato aaaa}

\item{Limite}{integer, Limitar la respuesta a un número de registros}
}
\description{
Función que hace un requerimiento de expedientes
}
