% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expedientes.R
\name{getExpedienteCabeza}
\alias{getExpedienteCabeza}
\title{Funcion que requiere el expediente cabeza de un expediente}
\usage{
getExpedienteCabeza(IdExpediente)
}
\arguments{
\item{IdExpediente}{integer ID de un expediente}
}
\value{
Un objeto de clase "tibble"
}
\description{
Funcion que requiere el expediente cabeza de un expediente
}
\examples{
prueba_cabeza <- getExpedienteCabeza(98985)

}
\author{
Martin Olmos, \email{molmos@itba.edu.ar}
}
