% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expedientes.R
\name{getExpMovimientos}
\alias{getExpMovimientos}
\title{Funcion para requerir los movimientos de un expediente}
\usage{
getExpMovimientos(IdExpediente)
}
\arguments{
\item{IdExpediente}{integer Numero de ID de un expediente}
}
\description{
Funcion para requerir los movimientos de un expediente
}
