% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expedientes.R
\name{getVotacionesExpediente}
\alias{getVotacionesExpediente}
\title{Funcion que requiere las votaciones de un expediente}
\usage{
getVotacionesExpediente(IdExpediente)
}
\arguments{
\item{IdExpediente}{integer ID de un expediente}
}
\description{
Funcion que requiere las votaciones de un expediente
}
