% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expedientes.R
\name{girosToDF}
\alias{girosToDF}
\title{Función que parsea un giro de XML a tibble}
\usage{
girosToDF(Giros)
}
\arguments{
\item{Giros}{giros de un expediente en formato XML}
}
\description{
Función que parsea un giro de XML a tibble
}
