% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expedientes.R
\name{giroToDF}
\alias{giroToDF}
\title{Función que parsea un giro de XML a tibble}
\usage{
giroToDF(Giro)
}
\arguments{
\item{Giro}{giro en formato XML}
}
\description{
Función que parsea un giro de XML a tibble
}
