% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expedientes.R
\name{moviToDF}
\alias{moviToDF}
\title{Función que parsea los movimientos de un expediente de XML a tibble}
\usage{
moviToDF(Movimientos)
}
\arguments{
\item{Movimientos}{movimientos de un expediente en formato XML}
}
\description{
Función que parsea los movimientos de un expediente de XML a tibble
}
