% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expedientes.R
\name{votacionesToDF}
\alias{votacionesToDF}
\title{Función que parsea las votaciones de un expediente de XML a tibble}
\usage{
votacionesToDF(Votaciones)
}
\arguments{
\item{Votaciones}{expediente cabeza de un expediente en formato XML}
}
\description{
Función que parsea las votaciones de un expediente de XML a tibble
}
