% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expedientes.R
\name{expedienteToDF}
\alias{expedienteToDF}
\title{Función que parsea un expediente de XML a tibble}
\usage{
expedienteToDF(Expediente = expediente)
}
\arguments{
\item{Expediente}{expediente en formato XML}
}
\description{
Función que parsea un expediente de XML a tibble
}
