% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expedientes.R
\name{getExpedienteCabeza}
\alias{getExpedienteCabeza}
\title{Funcion que requiere el expediente cabeza de un expediente}
\usage{
getExpedienteCabeza(IdExpediente)
}
\arguments{
\item{IdExpediente}{integer ID de un expediente}
}
\description{
Funcion que requiere el expediente cabeza de un expediente
}
