% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expedientes.R
\name{getExpedienteGiros}
\alias{getExpedienteGiros}
\title{Funcion para requerir los giros de un expediente}
\usage{
getExpedienteGiros(IdExpediente)
}
\arguments{
\item{IdExpediente}{Integer Numero de ID del expediente}
}
\description{
Funcion para requerir los giros de un expediente
}
