% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expedientes.R
\name{getExpedienteGiros}
\alias{getExpedienteGiros}
\title{Funcion para requerir los giros de un expediente}
\usage{
getExpedienteGiros(IdExpediente)
}
\arguments{
\item{IdExpediente}{Integer Numero de ID del expediente}
}
\value{
Un objeto de clase "tibble"
}
\description{
Funcion para requerir los giros de un expediente
}
\examples{
prueba_giros <- getExpedienteGiros(75640)

}
\author{
Martin Olmos, \email{molmos@itba.edu.ar}
}
