% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expedientes.R
\name{getExpMovimientos}
\alias{getExpMovimientos}
\title{Funcion para requerir los movimientos de un expediente}
\usage{
getExpMovimientos(IdExpediente)
}
\arguments{
\item{IdExpediente}{integer Numero de ID de un expediente}
}
\value{
Un objeto de clase "tibble"
}
\description{
Funcion para requerir los movimientos de un expediente
}
\examples{
prueba_movs <- getExpMovimientos(75640)

}
\author{
Martin Olmos, \email{molmos@itba.edu.ar}
}
